requirements.txt
biopython>=1.80
scikit-learn>=1.3.0
xgboost>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
joblib>=1.3.0
